requirements.txt
pandas
matplotlib
jupyter
openpyxl
requests
streamlit
altair
python-dotenv
pydantic>=2.0.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
sqlalchemy>=2.0.0
tenacity>=8.2.0